Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    float g;
+    int g;
     scanf("%f", &g);
-    printf("%d", g);
+    printf("%f", g);
     return 0;
 
 }
